Enforce Flow permissions and theme flow UI
