+++ internal/plugins/bundled/registry.json
@@ -858,7 +858,7 @@
         }
       }
     },
-    "entityPanels": [],
+    "entityPanels": ["Flow"],
     "actionTypes": []
   },
   {
