+++ internal/plugins/errors.go
@@ -1,0 +1,5 @@
+package plugins
+
+import "errors"
+
+var ErrPluginNotInstalled = errors.New("plugin not installed")
